requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pytz==2021.3
redis==3.5.3
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
